+++ src/dds.imageio/squish/squish-simd.h
